+++ src/Models/Guild.php
@@ -59,8 +59,6 @@ namespace CharlotteDunois\Yasmin\Models;
  * @property \CharlotteDunois\Yasmin\Interfaces\GuildChannelInterface|null   $widgetChannel                The guild's widget channel, or null.
  */
 class Guild extends ClientBase {
-    use \CharlotteDunois\Yasmin\Traits\HasImageTrait;
-
     /**
      * Guild default message notifications.
      * @var array
@@ -977,7 +975,7 @@ class Guild extends ClientBase {
      * @throws \InvalidArgumentException If $size is not a power of 2
      */
     function getBannerURL(?int $size = null, string $format = 'png') {
-        if(!$this->isPowerOfTwo($size)) {
+        if(!\CharlotteDunois\Yasmin\Utils\ImageHelpers::isPowerOfTwo($size)) {
             throw new \InvalidArgumentException('Invalid size "'.$size.'", expected any powers of 2');
         }
         
@@ -996,12 +994,12 @@ class Guild extends ClientBase {
      * @throws \InvalidArgumentException If $size is not a power of 2
      */
     function getIconURL(?int $size = null, string $format = '') {
-        if(!$this->isPowerOfTwo($size)) {
+        if(!\CharlotteDunois\Yasmin\Utils\ImageHelpers::isPowerOfTwo($size)) {
             throw new \InvalidArgumentException('Invalid size "'.$size.'", expected any powers of 2');
         }
 
         if(empty($format)) {
-            $format = $this->getImageExtension($this->icon);
+            $format = \CharlotteDunois\Yasmin\Utils\ImageHelpers::getImageExtension($this->icon);
         }
         
         if($this->icon !== null) {
