+++ src/Utils/ImageHelpers.php
@@ -7,24 +7,19 @@
  * License: https://github.com/CharlotteDunois/Yasmin/blob/master/LICENSE
 */
 
-namespace CharlotteDunois\Yasmin\Traits;
+namespace CharlotteDunois\Yasmin\Utils;
 
 /**
- * Whatever has this trait has an image (icon, avatar, etc.)
- *
- * The reason why there's no $image property is because certain classes can have
- * multiple images (like splash + icon for Guild, PartialGuild)
+ * Image Helper methods
  */
-trait HasImageTrait
-{
+class ImageHelpers {
     /**
      * Decides the default extension for an image
      *
      * @param  string $image Image hash
      * @return string        Depending on if animated or not, 'gif' or 'png'
      */
-    protected function getImageExtension(string $image) : string
-    {
+    static function getImageExtension(string $image) : string {
         return (\strpos($image, 'a_') === 0 ? 'gif' : 'png');
     }
 
@@ -34,8 +29,7 @@ trait HasImageTrait
      * @param  int|null $size The size we want to check
      * @return boolean        Whether it's a power of two
      */
-    protected function isPowerOfTwo(?int $size) : bool
-    {
+    static function isPowerOfTwo(?int $size) : bool {
         return $size === null || !($size & ($size - 1));
     }
 }
